theme moved out from within HomeHeading.jsx to its own /assets/theme.js file.
